Add missing migration
